+++ tests/testthat/test-t05_load_binpaths.R
@@ -18,7 +18,7 @@ test_that("Load empty binpaths", {
   binpaths_file <- normalize_path(file.path(
     if (dir.exists("~/.sen2r")) {"~/.sen2r"} else {tempdir()},
     "paths.json"
-  ))
+  ), mustWork = FALSE)
   if (file.exists(binpaths_file)) {file.remove(binpaths_file)}
   binpaths_0 <- load_binpaths()
   expect_is(binpaths_0, "list")
@@ -40,14 +40,14 @@ if (Sys.info()["sysname"] != "Windows") {
   test_that("Load GDAL", {
     binpaths_2 <- load_binpaths(c("gdal"))
     expect_is(binpaths_2, "list")
-    expect_length(binpaths_2, 9)
+    expect_length(binpaths_2, 3)
     expect_equal(binpaths_2$gdalinfo, normalize_path(Sys.which("gdalinfo")))
     expect_equal(basename(attr(binpaths_2, "path")), "paths.json")
   })
   test_that("Load aria2", {
     binpaths_3 <- load_binpaths("aria2")
     expect_is(binpaths_3, "list")
-    expect_length(binpaths_3, 10)
+    expect_length(binpaths_3, 4)
     expect_equal(binpaths_3$aria2, normalize_path(Sys.which("aria2c")))
     expect_equal(basename(attr(binpaths_3, "path")), "paths.json")
   })